refactor: PostMyBookCommand を作成
